% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_admixture_labels}
\alias{add_admixture_labels}
\title{Add admixture proprortions or admixture variables to a plot.}
\usage{
add_admixture_labels(plt, mapping, ...)
}
\arguments{
\item{plt}{A graph plot.}

\item{mapping}{A mapping from edge ids to the labels to add.}

\item{...}{Parameters that will be forwarded to \code{\link[ggraph:geom_edge_link]{ggraph::geom_edge_link()}}.}
}
\value{
An updated plot.
}
\description{
Add admixture proprortions or admixture variables to a plot.
}
