% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{dot_parse_graph}
\alias{dot_parse_graph}
\title{Read a list of edges from a 'dot' description.}
\usage{
dot_parse_graph(text)
}
\arguments{
\item{text}{A string that contains a graph in the \code{dot} format from
GraphViz.}
}
\value{
A linked list containing all the edges in the GraphViz graph.
Each edge has an associated label. We can use this to import
admixture proportions estimated by qpGraph.
}
\description{
Read a list of edges from a 'dot' description.
}
