% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_graph_plot}
\alias{make_graph_plot}
\title{Construct a graph plot object.}
\usage{
make_graph_plot(graph, ...)
}
\arguments{
\item{graph}{An admixture/drift graph.}

\item{...}{Parameters passed to \code{\link[ggraph:geom_edge_link]{ggraph::geom_edge_link()}} that you can
use to customise the edges.}
}
\value{
A graph object
}
\description{
This object can be modified using \code{ggraph} and \code{ggplot2} functions.
}
