% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{parse_dot}
\alias{parse_dot}
\title{Parse a graph in GraphViz' .dot format}
\usage{
parse_dot(text)
}
\arguments{
\item{text}{Text containing the graph description.}
}
\description{
Parse a graph in GraphViz' .dot format
}
