% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgraph.R
\name{parse_qpgraph}
\alias{parse_qpgraph}
\title{Parse a qpGraph text.}
\usage{
parse_qpgraph(text)
}
\arguments{
\item{text}{The text description of the graph}
}
\value{
An admixture graph
}
\description{
Parse a qpGraph text.
}
