% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{read_dot}
\alias{read_dot}
\title{Import a dot file into an admixturegraph object}
\usage{
read_dot(text)
}
\arguments{
\item{text}{Text containing the graph description.}
}
\value{
An admixturegraph object
}
\description{
Import a dot file into an admixturegraph object
}
