% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{show_inner_node_labels}
\alias{show_inner_node_labels}
\title{Add inner nodes labels}
\usage{
show_inner_node_labels(plt, ...)
}
\arguments{
\item{plt}{Plot object created by \code{\link[=make_graph_plot]{make_graph_plot()}}.}
}
\value{
Updated plotting object
}
\description{
Add inner nodes labels
}
