% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{show_leaf_labels}
\alias{show_leaf_labels}
\title{Add leaf labels.}
\usage{
show_leaf_labels(plt, ...)
}
\arguments{
\item{plt}{Plot object created by \code{\link[=make_graph_plot]{make_graph_plot()}}.}

\item{...}{Parameter forwarded to \code{\link[ggraph:geom_node_text]{ggraph::geom_node_text()}}.}
}
\value{
Updated plotting object
}
\description{
This is a wrapper around \code{\link[ggraph:geom_node_text]{ggraph::geom_node_text()}} and you can customise
the plotting through \code{...}
}
