% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_graph_plot}
\alias{make_graph_plot}
\title{Construct a graph plot object.}
\usage{
make_graph_plot(graph, ...)
}
\arguments{
\item{graph}{An admixture/drift graph.}

\item{...}{Parameters that will be forwarded to \code{\link[ggraph:geom_edge_link]{ggraph::geom_edge_link()}}.}
}
\value{
A graph plot object.
}
\description{
Parameters in \code{...} will be forwarded to \code{\link[ggraph:geom_edge_link]{ggraph::geom_edge_link()}}.
The returned object can be modified using \code{ggraph} and \code{ggplot2} functions.
}
