% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_graph_plot}
\alias{make_graph_plot}
\title{Construct a graph plot object.}
\usage{
make_graph_plot(graph, edge_colour = "darkblue", leaf_label_angle = 0,
  leaf_label_size = 4)
}
\arguments{
\item{graph}{An admixture/drift graph.}

\item{edge_colour}{The colour to use for the graph edges.}

\item{leaf_label_angle}{Angle to display leaf nodes in.
Useful if they would otherwise overlap.}
}
\value{
A graph object
}
\description{
This object can be modified using \code{ggraph} and \code{ggplot2} functions.
}
