% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{show_leaf_labels}
\alias{show_leaf_labels}
\title{Add ileaf labels}
\usage{
show_leaf_labels(plt, ...)
}
\arguments{
\item{plt}{Plot object created by \code{\link[=make_graph_plot]{make_graph_plot()}}.}
}
\value{
Updated plotting object
}
\description{
Add ileaf labels
}
